ocn: gestures: modify long squeeze activation behaviour
An issue that we had was that the user had no feedback as to how long to squeeze down to activate the long squeeze gesture.
Rework this so that the long squeeze gesture activates if the edge sensors are squeezed down for a certain period of time instead of the instance the squeeze is released.

Change-Id: Ic910c2876bf8b280e7cb7de8c0f9ede895934895