ocn: gestures: revamp for the use of the HTC Edge Gesture Sensor
The HTC Edge Sensor currently used does not work well in deep sleep and
the vast majority of sensor events sent are received too late, rendering
Edge Sense completely useless in deep sleep.

On the other hand, the HTC Edge Gesture Sensor is able to send events
even in deep sleep, so listen to sensorevents coming from this sensor
and allow it to support the HTC Edge Sensor by acquiring a wakelock when
the Edge sensors are squeezed while the phone is in deep sleep.

Stock EdgeSensorService was partly used as a reference guiding some changes.

Change-Id: I0b4ed260b2a985069db6d7cc570d002fbef6df58