ocn: SqueezeGestureService: revamp squeeze gesture handling yet again
By returning to stock and observing squeeze gesture behaviour as well
as decompiling the EdgeSensorService apk. It can be seen that the squeeze
gesture implementation by HTC is completely different from ours.

The use of the HtcEdgeSensor sensor here is completely incorrect - HTC
merely used this sensor to display the squeeze effects around where the
sensors are located on the screen. The squeeze gestures themselves were
completely handled by HtcEdgeGestureSensor. Modify the code here to do
exactly this and totally remove HtcEdgeSensor as it is not useful for us.

Keep the long squeeze gesture activation behaviour, as without visual
feedback, it is difficult to determine how long to hold squeeze down for.
This could be workarounded by forcing a vibration when the long squeeze
duration has elapsed, but vibration can be disabled so this would be useless.

Change-Id: Ie589339c340bddec364eebcc8261d77f3b7d72f1